requirements/deploy.txt
Django==1.10.7
django-bulk-update
django-ckeditor
django-filter
django-mail-templated
django-mathfilters
django-mptt
django-nested-admin
Pillow
psycopg2
pytz
lxml
xmltodict
